requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
slowapi==0.1.9
pydantic==2.7.4
python-dotenv==1.0.0
openai==1.52.0
psycopg2-binary==2.9.9
pgvector==0.3.2
langgraph==0.2.28
langchain==0.3.7
langchain-community==0.3.7
pandas==2.1.3
requests==2.31.0
httpx==0.27.2
duckduckgo-search==6.2.6

# Testing
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27.2